+++ src/DevilDaggersInfo.Core.Mod/Structs/Vertex.cs
@@ -1,15 +1,3 @@
 namespace DevilDaggersInfo.Core.Mod.Structs;
 
-public struct Vertex
-{
-	public Vertex(Vector3 position, Vector3 normal, Vector2 texCoord)
-	{
-		Position = position;
-		Normal = normal;
-		TexCoord = texCoord;
-	}
-
-	public Vector3 Position { get; }
-	public Vector3 Normal { get; }
-	public Vector2 TexCoord { get; }
-}
+public readonly record struct Vertex(Vector3 Position, Vector3 Normal, Vector2 TexCoord);