+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
-- `Spawn` is now a record struct.
+- `Spawn`, `Vertex`, and `VertexReference` are now record structs.
 - `SpanWrite` is now internal. This type wasn't supposed to be public.
 
 ### Removed